LPD-78225 Fix. Export findActions utility from the FDS. Use it from custom cell renderers
