+++ modules/apps/frontend-data-set/frontend-data-set-web/src/main/resources/META-INF/resources/index.ts
@@ -6,6 +6,7 @@
 export {default as FrontendDataSet} from './FrontendDataSet';
 export {default as FrontendDataSetContext} from './FrontendDataSetContext';
 export {default as DateRenderer} from './cell_renderers/DateRenderer';
+
 export {default as DateTimeRenderer} from './cell_renderers/DateTimeRenderer';
 
 export {INTERNAL_CELL_RENDERERS as FDS_INTERNAL_CELL_RENDERERS} from './cell_renderers/InternalCellRenderer';
@@ -31,6 +32,8 @@ export {
 } from './inline_notification/InlineNotification';
 
 export {ACTION_ITEM_TARGETS} from './utils/actionItems/constants';
+
+export {default as findAction} from './utils/actionItems/findAction';
 
 export {replaceTokens} from './utils/actionItems/formatActionURL';
